Refactor use of some package-level imports from theano.link
