+++ tests/link/test_cc.py
@@ -6,7 +6,7 @@ from theano.gof import fg
 from theano.gof.graph import Apply, Constant, Variable
 from theano.gof.op import Op
 from theano.gof.type import Type
-from theano.link import PerformLinker
+from theano.link.basic import PerformLinker
 from theano.link.c.cc import CLinker, DualLinker, OpWiseCLinker
