+++ theano/compile/monitormode.py
@@ -2,8 +2,10 @@
 # original author, and re-licensed under Theano's license.
 import numpy as np
 
-import theano
 from theano.compile.mode import Mode
+from theano.configdefaults import config
+from theano.link.basic import WrapLinkerMany
+from theano.link.c.cc import OpWiseCLinker
 
 
 class MonitorMode(Mode):
@@ -40,11 +42,9 @@ class MonitorMode(Mode):
     def __init__(self, pre_func=None, post_func=None, optimizer="default", linker=None):
         self.pre_func = pre_func
         self.post_func = post_func
-        wrap_linker = theano.link.WrapLinkerMany(
-            [theano.link.c.cc.OpWiseCLinker()], [self.eval]
-        )
+        wrap_linker = WrapLinkerMany([OpWiseCLinker()], [self.eval])
         if optimizer == "default":
-            optimizer = theano.config.optimizer
+            optimizer = config.optimizer
         if linker is not None and not isinstance(linker.mode, MonitorMode):
             raise Exception(
                 "MonitorMode can only use its own linker! You "
@@ -95,13 +95,15 @@ class MonitorMode(Mode):
 
 
 def detect_nan(fgraph, i, node, fn):
+    from theano.printing import debugprint
+
     for output in fn.outputs:
         if (
             not isinstance(output[0], np.random.RandomState)
             and np.isnan(output[0]).any()
         ):
             print("*** NaN detected ***")
-            theano.printing.debugprint(node)
+            debugprint(node)
             print("Inputs : %s" % [input[0] for input in fn.inputs])
             print("Outputs: %s" % [output[0] for output in fn.outputs])
             break
