+++ theano/scan/scan_perform.pyx
@@ -61,7 +61,8 @@ cimport numpy
 import copy
 import time
 
-from theano import gof, link
+from theano import gof
+from theano.link.utils import raise_with_op
 
 
 def get_version():
@@ -405,7 +406,7 @@ def perform(
                 # done by raise_with_op is not implemented in C.
                 if hasattr(fn, 'thunks'):
                     # For the CVM
-                    link.raise_with_op(fn.maker.fgraph,
+                    raise_with_op(fn.maker.fgraph,
                                            fn.nodes[fn.position_of_error],
                                            fn.thunks[fn.position_of_error])
                 else:
@@ -413,7 +414,7 @@ def perform(
                     # We don't have access from python to all the
                     # temps values So for now, we just don't print
                     # the extra shapes/strides info
-                    link.raise_with_op(fn.maker.fgraph, fn.nodes[fn.position_of_error])
+                    raise_with_op(fn.maker.fgraph, fn.nodes[fn.position_of_error])
             else:
                 # old-style linkers raise their own exceptions
                 raise
